[PWA] FetchInstallInfoFromInstallUrlCommand uses install info creation
This CL migrates the FetchInstallInfoFromInstallUrlCommand to now use
the newly created ManifestToWebAppInstallInfoJob to generate the install
info from a manifest. To do this, the following changes were required:

1. The job now supports the setting of whether favicons need to be
skipped or not from the caller who invokes it via the options struct.
2. The command now performs 2 paths instead of the consolidated one path
before. One path runs in the existence of a manifest, and moves over
fields obtained from the web page metadata into the WebAppInstallInfo
created by the job, while the other path follows the old code path
of what happens when a manifest is not found.

Tested using the unit-tests for the command, as well as the sub apps
install service and PWA devtools protocol browser tests who both
use this command to ensure no regressions in behavior.

Bug: 427565907
Change-Id: Ib860ed80dee0ad0d6e24c00f1e0ec951d6c73e8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681208
Reviewed-by: May Siem <msiem@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480019}