Add Trial to keep Persistent Histograms source on FRE on iOS.
RegisterSource runs on startup, and deletes the sources when metrics are
disabled. This is a problem on First Run, since metrics will be
disabled, at least until the client finishes it. Therefore, the
persistent histograms that are not sent to the server will not persist
when closing Chrome.

Add experiment to not delete those sources when it's the First Run. In
the case metrics are still disabled on subsequent Chrome launches, the
sources will be deleted following the current behavior. We only want to
start the experiment for users that are in the First Run. They'll be
assigned to the same group in subsequent Chrome launches.

For android webview, the parameter `is_first_run` is always set to false
to avoid changing its original behaviour.

For now the experiment will only be enabled on iOS. The rest of the
platforms will not participate in the experiment (is_first_run=false).

Bug: 407991309

Change-Id: I8547aed8603608156e02f9ccefe05a951d9dff77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6620860
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Commit-Queue: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1480044}