Add UiEvent types along with a DebugString method
These are the objects that will be propagated through the UI layer.
They are intentionally lower dependency than ToolRequests, and one
ToolRequest may be converted into one or more UiEvents.

In this CL we introduce the first three, which we think are the
highest priority.  Some very rough notes on how these objects will
be used in the context of execution are here:
https://docs.google.com/document/d/1n4c5f61-1MdRpxSPtnsijruJr93Orq2Gn2YJdbPV6Qk/r/0-o2IIbtQtr2csqLXyu9IYUA/edit?tab=t.0

Bug: 425784083
Change-Id: Ib308835464a99ba146f40ce6d0d5ff3f9378eecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681472
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Kevin Graney <kmg@google.com>
Cr-Commit-Position: refs/heads/main@{#1480071}