[SxS] Add split view regression tests for Extensions API
The goal of these tests are to ensure that existing tabs API methods
won't break extensions when Split View is rolled out. The tested
functions include move, duplicate, remove, and discard. Our general
strategy is to fail gracefully and remove splits when operations are
inconsistent. For discard, the behavior is that the function will fail
if a visible but inactive tab is discarded.

Bug: 406793241
Change-Id: Idbc09f6250e5afe31267897d5048016a8b2d84d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6664920
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Reviewed-by: Solomon Kinard <solomonkinard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480080}