[Masonry] Fix row-auto-placement-min-content ref file
The reference file for row-auto-placement-min-content was not passing
test runs given that the reference was not displaying the correct
container width; now the background has the width of if the items were
laid out with min-content, and the height as if the items were laid out
with max-content. This is the correct behavior because in masonry, the
width of the container is calculated as if the items are all laid out
with min-content, but the height ends up being shorter since during the
layout pass we default to fit-content behavior, and the items are laid
out in max-content width.

Bug: 343257585
Change-Id: If90e0b24fc55c507c891d0c4ee58e5120b2759ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6675055
Commit-Queue: Celeste Pan <celestepan@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1480088}