extensions: Port DNS API and test to desktop Android
The API itself is straightforward and only requires enabling the
dns permission for desktop_android in _permission_features.json.

Convert the test from a ShellApiTest to an ExtensionApiTest. We don't
have ShellApiTest on desktop Android and we've decided not to try to
port it, because running the tests as Chrome browser_tests is a better
simulation of production.

Move dns_apitest.cc into //chrome so it can be an ExtensionApiTest.
Move the test data as well.

Bug: 405219601
Change-Id: Ia22ac9010fcec8ad2bb1232a7a5eb9c46777abe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6680793
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Tim <tjudkins@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Zhengzheng Liu <zhzhliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480097}