Add support for same document navigation.
Safe-browsing lookups do not consider fragment URLs distinct for the
same host name and path. So, if watermark.com is cached, navigating to
watermark.com#page would not trigger a lookup and retrieve the cached
verdict for watermark.com. A parameter was therefore introduced in
https://crrev.com/c/6637995, to bypass the SB cache in that case. When
finch is disabled, the lookup defaults to the original behaviour, since
StartMaybeCachedLookup(use_cache=true) is equivalent to StartLookup().

Bug: 421458310
Change-Id: Ie89fe396b575f64f93822c1efc3886e69251a9c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6675903
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nasser Al-shawwa <alshawwa@chromium.org>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480108}