pgo: Add dynamic PGO profile path detection for Android
This change introduces a new "auto" option for the
`--android-device-path` and `--fetch-data-path-device` flags. When
"auto" is used, the script will dynamically detect the PGO profile path
on the device, accommodating for different user IDs in the path
structure.

This avoids the need to hardcode the user ID in the path, making PGO
profile generation more robust on Android.

Bug: 418315677
Change-Id: I7e30825bdb637670575d9c734ead4cdbfc0cff30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6678302
Commit-Queue: Darren Wu <darrenwu@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/main@{#1480119}