[PDF] Replace UNSAFE_TODO in PDFiumEngine::GetRegion()
Replace the UNSAFE_TODO with UNSAFE_BUFFERS in
PDFiumEngine::GetRegion(). The UNSAFE_BUFFERS wrapper cannot be removed
since the block still involves creating a buffer using a pointer and a
size value. Instead, leave a comment explaining why it is still safe,
and add a TODO to switch to SkSpan when it becomes part of the public
Skia API.

Change-Id: I8c6a67fa2ce852c586109e72e9e1b61de04f330b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681110
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480142}