[PDF Ink Signatures] Add a text caret in the PDF content area
Introduce a text caret in the PDF content. Users will be able to use
this to scroll around the PDF and select text with only a keyboard. Add
a PdfCaret class that PDFiumEngine will own to handle drawing the
blinking caret. In the future, PdfCaret will also handle mouse events
and key events. For now, PdfCaret only shows a blinking caret on the
first character of a PDF text page.

Because caret browsing mode is being implemented to provide keyboard
support to the Ink2 text highlight feature, it will only be enabled when
Ink2 text highlighting is enabled. In the future, caret browsing can be
rolled out as its own feature in the PDF viewer.

Bug: 427139499, 427778119
Change-Id: Ib735be194ee40ef9d115d3a46933f76604905764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6664049
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480146}