Add a basic typemap for mojom.ACMatchClassification
This reduces binary size by about 700 bytes on Android: while it's not a
lot, it is a demonstration of the inefficiency that type converters
introduce, since they effectively force types to be serialized twice:
once when passing through the converter, and once when being written out
to the wire.

This CL also changes an assignment to a vector of ACMatchClassification
from:
  match.contents_class = {{}};
to:
  match.contents_class = {};

Mojo deserialization requires types to be default constructible; adding
a default constructor to ACMatchClassification changes the meaning of
the original statement (!!) from  "construct an a vector from an empty
initializer list" to "construct an initializer list with a single
default constructed ACMatchClassification element".

Bug: 428021766
Change-Id: Ic88af22f09d58b4f0760e0e880259e828fe1c575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6682802
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480178}