ownership: support crypto/keypair API in MockOwnerKeyUtil
This allows clients to use either the old crypto/rsa_private_key API or
the new crypto/keypair API with this utility class, which will make it
easier for clients to migrate to the new API if they want to. The
behavior internally is identical, but the new API passes keys around as
values rather than as heap allocations (the actual key internally is
always heap-allocated in either API, but keypair::PrivateKey is a value
type) and takes/returns values directly rather than using out params.

Bug: 425863216
Change-Id: Ied2a2b208b89c85345a2e681d2247ef757e651f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681753
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480341}