Add comparison operator to ConnectionEndpointMetadata
This is a preparation for coalescing QuicSessionAttempts for a
destination in the HEv3 code path. Add comparison operator to
ConnectionEndpointMetadata so that we can use it as a key of a set.
I plan to use the class as a part of a key that coalesces
QuicSessionAttempts.

Bug: 416364483
Change-Id: Ia3578a7e6e2ce5dde3181831b33f011244f906f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6682758
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480385}