DsePreload2: Delay a check for limit exceeded
This CL delays a check for limit exceeded of prefetch, so that we record
`SearchPreloadSignalResult::kPrefetchLimitExceeded` only if the item
is suggested to prefetch. Before this patch, it is recorded whenever
the Omnibox is updated after limit exceeded, even if the item is not
suggested.

Bug: 394213503, 424142992
Change-Id: I4722ffbcb7b5a678e02bf593ebb8890d9ec219ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6657762
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Ken Okada <kenoss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480417}