[autofill_ml] Set up mojo bindings for autofill-ml-internals WebUI
This will allow us to pass arbitrary structured information from the
browser to the frontend.

The structure of the mojo object will change in the future when we pass
more data (model output probabilities, tokenized representations of
strings, and so on).

For now, a placeholder `MLPredictionLog` is created in C++ in the
`SetPage` method of the page handler. It is then passed to the frontend
and displayed. This provides a nice sanity check that the IPC is
working.

Change-Id: I3f6c2e14dc1a72653a4a866d5e4e40d9f9cfbf5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6632681
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Teo Lamort de Gail <lamort@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480444}