[Unowned Data] Transition to TypedIdentifier
This simplifies and transfers UnownedUserDataCollection to using
TypedIdentifier, which provides some nice guarantees. In order to
restrict access to the host, it does not remove PassKeys, despite that
being part of the original proposal (if we decide to do that, it will
be in another CL).

Follow-up will be factories and better test substitution and injection,
also as per the proposal (see attached bug).

Bug: 427458853
Change-Id: I2e63902355721d7cd78484fa529e05d83fdd12fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6668552
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480511}