[Unowned Data] Enable injection for testing.
Creates the concept of a UserDataFactory which allows Features-type
objects to enable injection of alternate objects (such as mocks and
fakes) in tests.

An toy example is provided in the accompanying tests.

See the document linked in the bug for more information.

Follow-up: Add factories to BrowserWindowFeatures, TabFeatures; switch
to factory overrides in "real" tests.

Bug: 427458853
Change-Id: Iaec22c3d66454043db4bb68dec04d920f1b9bfe2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6667467
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480512}