[ACT] Minor readability updates and fixes to UB logic
- Change `RegisterOnUserTriggeredReloadingActionCallback` back to
`RegisterOnUserClosedContentViewCallback`: it was originally renamed in
crrev.com/c/6437441, but looking at this logic the rename doesn't make
sense as it's a callback for the user closing the view, which doesn't
mean triggering a reload (esp for ACT)

- Refactor logic in `OnUserClosedContentView` to specifically match
what is needed by `RegisterOnUserClosedContentViewCallback` and move
ACT-specific logic into a separate function as the ACT logic is never
needed by the callback (i.e. it only ever closes the view for ACT). This
reduces ~20 lines from the view controller and gets rid of unneeded
nesting + indirection

- Replace `SetSpinnerVisible` with `GetTrackingProtectionsButton`:
having a function called `SetSpinnerVisible` in the top-level content
view API is strange since a spinner never shows in the top-level content
view, it's nested in the button (and therefore should be set via the
button view API)

- Logic fix to icon view to correctly set tooltip text in the "resumed"
case and have it match the icon label

Change-Id: I5007699736b7bba32314ca3f8574a899426dec22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681394
Auto-Submit: Fiona Macintosh <fmacintosh@google.com>
Reviewed-by: Michelle Abreo <michelleabreo@google.com>
Commit-Queue: Michelle Abreo <michelleabreo@google.com>
Cr-Commit-Position: refs/heads/main@{#1480554}