Disable pull-to-refresh consistently if overflow_y_hidden
In the linked bug, pull-to-refresh sometimes worked and sometimes
didn't work on pages with overflow: hidden on <html> or <body>.
This appears to be because there is a race between
OverscrollRefresh::OnOverscrolled and
OverscrollRefresh::WillHandleScrollUpdate after a scroll has begun.
When OnOverscrolled happens first, pull-to-refresh will be allowed,
even though we know that OverscrollRefresh::overflow_y_hidden is true.

This patch addresses this by performing MaybeDisableScrollConsumption
a little earlier in OnOverscrolled so we do not discard the resulting
change of state of OverscrollRefresh::scroll_consumption_state_.

Bug: 428661784
Change-Id: I082050678f33227f3268b7894354e2a6e45bef25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6688729
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480558}