[cc] Fix dangling ptr in LayerTreeTestLayerTreeFrameSinkClient
The LayerTreeTestLayerTreeFrameSinkClient pointer to the
TaskRunnerProvider was a provider owned by the LayerTreeHost. But the
client was also pointed at by the TestLayerTreeFrameSink and
TestCompositorFrameSinkSupport so we must clean those up before cleaning
up the client or the LayerTreeHost. This change implements this inside
of DestroyLayerTreeHost.

To tear down the sink, we must set the visibility to false to avoid
triggering a DCHECK when releasing the sink. Originally this caused some
tests to fail because they were validating visibility state in their
AfterTest overrides, assuming that it was always visible. But the
previous changes to move AfterTest to before destroying the LTH, and
combining it with CleanupBeforeDestroy, provides a consistent shutdown
path so that AfterTest is the state immediately after EndTest and not
affected by any side-effects of destroying LTH.

Additionally, TestLayerTreeFrameSink::DetachFromClient could suffer from
unexpected reentrancy, leading to a crash in DidNotProduceFrame when
trying to access support_ after releasing it. This CL includes a check
to prevent that.

Bug: 40212619, 426247403
Change-Id: I92abaf447a70bb7a0ccced9bb720da8370e8b141
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6669311
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1480585}