Forward keyboard events from XrHostActivity
Adds an override for `dispatchKeyEvent` in XrHostActivity to ensure that
keyboard events are still routed to the main activity and the underlying
page responsible for the WebXR content. This allows them to have key
event listeners that they can use to trigger different behavior.

Bug: 427502710
Change-Id: I0d3b9944795391b4973168d6c7260b7379dea777
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681350
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480610}