Fix HTTPS 500 errors pages shows not secure
When navigate to a server getting internal error response like code 500,
if the server doesn't provide an error page, browser will render a
default error page. In this case, info bubble in the omnibox shows as
"not secure". This is misleading, because address bar shows the
server/url, if it is secure with certificates, it should be regarded as
secure. And the error page is a local static page, it should be regarded
as secure. So, pages with valid TLS certificates chain need to be marked
as Secure, irrespective of the HTTP response code.

The issue is caused by missing ssl info. Even the server may own a valid
TLS certificate, when the error happens, navigation goes to the cancel
process and the request fail process, which will unintentionally erase
the ssl info.

This CL will add ERR_HTTP_RESPONSE_CODE_FAILURE check to deal with ssl
info update steps, during the request fail process. This will ensure the
ssl certificate is preserved. So, the ominbox info bubble can show the
above case as secure and show the correct certificate info for the
server even it's a default error page.

Bug: 41340251
Change-Id: I80c3a001c9aa1a3e3c80ab3b907c6c9e8541edd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6466824
Reviewed-by: Liang Zhao <lzhao@microsoft.com>
Commit-Queue: Eain Chen <eainchen@microsoft.com>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480625}