ChromeVox: Use SRE in manifest v3
This CL imports SRE into manifest v3 ChromeVox. We mostly follow the
pattern established by the mv2 code in https://crrev.com/c/6534810,
but with the key difference of calling SRE from the offscreen
document. For context, we cannot run SRE in the service worker
context because the dependency contains various web API calls that
are not allowed in a service worker. As a result, we also need to wire
up the messaging between the service worker and the offscreen
document to handle the calls to SRE.move() and SRE.walk().

Currently ChromeVox mv3 tests are not enabled in CQ. However, I have
uploaded a downstream change [1] that enables and runs the mv3
tests in CQ, which show no failures for any of the math-related tests.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6684988

Bug: 388867840
Change-Id: I3f540e483b5fb1f08109c30ea99877bd3849c071
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681310
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480628}