[AriaNotify] Ensure we fire the UIA event on a control element.
It is common for authors to fire events on the body, but since chromium
often doesn't consider the body to be in the UIA Content/Control View,
these events are filtered out by ATs. This change ensures that we bubble
up the event to the nearest ancestor that is in the Content/Control
View. From my understanding and observations, the document will always
be in the Content/Control tree, so there should always be an ancestor to
fire this on.

Bug: 326277796
Change-Id: I6c0fc0348a7c66c084230704f536f89ccf39f8b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6676264
Auto-Submit: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1480644}