[CLAMS] Let ClientSideDetectionIntelligentScanDelegateDesktop listen to pref changes directly.
The previous plan was to let ClientSideDetectionService call
Start/StopListeningToOnDeviceModelUpdate when the pref changes. However,
that will cause a circular dependency error. If the service class calls
the delegate class, it needs to depend on the client_side_detection
target which includes the definition of the delegate. However, the
client_side_detection target already depends on
client_side_detection_service, causing circular dependency.

To fix it, the delegate will listen to pref changes by itself. Then the
service class doesn't need to call the delegate.

Other than listening to the change in enhanced protection, the service
class also checks kClientSideDetectionBrandAndIntentForScamDetection and
kClientSideDetectionLlamaForcedTriggerInfoForScamDetection. It also
checks whether client_side_phishing_model_ is null, which only happens
when the killswitch is enabled. These three flags are all checked in the
delegate as well.

Bug: 424104358
Change-Id: I716d5ab5416e94af4dbe1511adefd9d94570cd7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6682217
Reviewed-by: Seung Jae Lim <andysjlim@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480674}