[Unowned Data] Remove test factory for BrowserWindowFeatures
Since we can do injection and substitution via UnownedUserDataHost and
UserDataFactory, there's no reason to also be able to subclass
BrowserWindowFeatures for testing. (Also nobody was actually doing it.)

This also removes the necessity for a factory method that can be
overridden in tests; BrowserWindowFeatures can simply be directly
created.

Bug: 427458853
Change-Id: I8cedee53fdd0ae740fd5a2d75e9ac12f88341d83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6681570
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480690}